pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "foundry"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "brainsets",
    "hydra-core>=1.3.0",
    "hydra-submitit-launcher>=1.2.0",
    "ipykernel>=7.1.0",
    "lightning>=2.0.0",
    "torch>=2.9.1",
    "torch-brain",
    "temporaldata",
    "matplotlib>=3.10.8",
    "torchmetrics>=1.8.2",
    "wandb>=0.24.0",
    "auditorydecoding",
]

[dependency-groups]
dev = ["pre-commit>=4.5.1", "pytest>=8.0.0", "ruff>=0.14.14"]

[tool.uv.sources]
torch-brain = { git = "https://github.com/neuro-galaxy/torch_brain", branch = "milo/fast-sampler" }
brainsets = { git = "https://github.com/neuro-galaxy/brainsets", branch = "mila/dev-main" }
temporaldata = { git = "https://github.com/neuro-galaxy/temporaldata", branch = "main" }
auditorydecoding = { git = "https://github.com/Neurosoft-Bioelectronics/auditorydecoding", branch = "milo/temp-dep-devmain" }

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 80

[tool.ruff.lint.pydocstyle]
convention = "google"
